git-svn-id: svn://svn.code.sf.net/p/x10/code/branches/java-interop@22673 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src-java/x10/core/Byte.java
@@ -28,7 +28,7 @@ final public class Byte extends Number implements StructI, java.lang.Comparable<
     x10.lang.Arithmetic<Byte>, x10.lang.Bitwise<Byte>, x10.util.Ordered<Byte>
 {
     private static final long serialVersionUID = 1L;
-    private static final int _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(Byte.class);
+    private static final short _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(Byte.class);
     
     public static final x10.rtt.RuntimeType<?> $RTT = Types.BYTE;
     public x10.rtt.RuntimeType<?> $getRTT() {return $RTT;}
@@ -164,7 +164,7 @@ final public class Byte extends Number implements StructI, java.lang.Comparable<
         serializer.write($value);
     }
 
-    public int $_get_serialization_id() {
+    public short $_get_serialization_id() {
         return _serialization_id;
     }
