+++ x10.runtime/src-java/x10/core/UByte.java
@@ -27,7 +27,7 @@ final public class UByte extends Number implements StructI, java.lang.Comparable
     x10.lang.Arithmetic<UByte>, x10.lang.Bitwise<UByte>, x10.util.Ordered<UByte>
 {
     private static final long serialVersionUID = 1L;
-    private static final int _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(UByte.class);
+    private static final short _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(UByte.class);
     
     public static final x10.rtt.RuntimeType<?> $RTT = Types.UBYTE;
     public x10.rtt.RuntimeType<?> $getRTT() {return $RTT;}
@@ -118,7 +118,7 @@ final public class UByte extends Number implements StructI, java.lang.Comparable
         serializer.write($value);
     }
 
-    public int $_get_serialization_id() {
+    public short $_get_serialization_id() {
         return _serialization_id;
     }
